% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUN.R
\name{avg.arit}
\alias{avg.arit}
\title{arithmetic average of two models}
\usage{
avg.arit(mat1, mat2, label = train$popularity, iter = 101)
}
\arguments{
\item{mat1}{Matrix of first classifier class probabiliets.}

\item{mat2}{Matrix of second classifier class probabiliets.}

\item{label}{True class labels. Default train set popularity column.}

\item{iter}{Number of weight combinations, default 101 (0.01 increment)}
}
\value{
An ordered matrix with weight combinations and matching accuracies
}
\description{
This version is made for 5-label classification. Inputs are prediction
probability matricesof two classifiers.
}

