% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUN.R
\name{cv.h2odl}
\alias{cv.h2odl}
\title{cross validation neural network h2o package}
\usage{
cv.h2odl(train, epochs = 10)
}
\arguments{
\item{train}{Any train set without ID and URL}

\item{epochs}{Number of passes over the training dataset, default 10}
}
\value{
A list containing a vector of each fold accuracy and a mean.
}
\description{
five fold cross validation using h2o NN
}

