% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUN.R
\name{cv.nn}
\alias{cv.nn}
\title{cross validation neural network nnet package}
\usage{
cv.nn(train, size = 20, rounds = 500)
}
\arguments{
\item{train}{Any train set without ID and URL}

\item{size}{Number of units in hidden layer, default 20}

\item{rounds}{Number of rounds, default 500}
}
\value{
A list containing a vector of each fold accuracy and a mean.
}
\description{
five fold cross validation using nnet NN
}

