% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUN.R
\name{cv.rfmc}
\alias{cv.rfmc}
\title{cross validation random forest multicore}
\usage{
cv.rfmc(train, tree = 1000, core = 4)
}
\arguments{
\item{train}{Any train set without ID and URL}

\item{tree}{Number of trees, default 1000}

\item{core}{Number of cores, default 4}
}
\value{
A list containing a vector of each fold accuracy and a mean.
}
\description{
five fold cross validation using random forest
}

