% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUN.R
\name{cv.xgb}
\alias{cv.xgb}
\title{cross validation xgboost}
\usage{
cv.xgb(train, rounds = 250)
}
\arguments{
\item{train}{Any train set without ID and URL}

\item{rounds}{Number of rounds, default 250}
}
\value{
A list containing a vector of each fold accuracy and a mean.
}
\description{
five fold cross validation using xgboost
}

