% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUN.R
\name{fmeta.mab}
\alias{fmeta.mab}
\title{adaoost meta predictor (maboost package)}
\usage{
fmeta.mab(train, test = NULL, rounds = 250, csv = FALSE)
}
\arguments{
\item{train}{Train set without ID and URL}

\item{test}{Test set without ID and URL. If left empty then it the
functions trains on train with 5 fold cross validation}

\item{rounds}{Number of rounds per folds}

\item{csv}{Boolean, default FALSE. If TRUE, saves csv output.}
}
\value{
A data frame with 5 probabilites for each row
}
\description{
Creates probabilities for each class
}

