% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUN.R
\name{full.rf}
\alias{full.rf}
\title{DUN random forest model}
\usage{
full.rf(train, test)
}
\arguments{
\item{train}{full training set.}

\item{test}{full test set.}
}
\value{
A data frame with id and popularity columns
}
\description{
Creates team DUN submission file
}

