% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUN.R
\docType{data}
\name{test}
\alias{test}
\title{News popularity test data.}
\format{A data frame with 9644 rows and 61 variables}
\source{
\url{https://inclass.kaggle.com/c/predicting-online-news-popularity/data}
}
\usage{
test
}
\description{
A dataset containing the attributes of 9,644 articles.
}
\keyword{datasets}

