% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUN.R
\docType{data}
\name{train}
\alias{train}
\title{News popularity training data.}
\format{A data frame with 30000 rows and 62 variables}
\source{
\url{https://inclass.kaggle.com/c/predicting-online-news-popularity/data}
}
\usage{
train
}
\description{
A dataset containing the attributes of 30,000 articles.
}
\keyword{datasets}

