% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DUN.R
\name{word}
\alias{word}
\title{Word frequency}
\usage{
word(a)
}
\arguments{
\item{a}{String of a single word.}
}
\value{
A list with following elements: freqdiff, diff, freqtrain, freqtest
}
\description{
Gives word frequencies etc
}

